Update backup.js
Sorting:
fileData.sort((a, b) => new Date(b.datecreated) - new Date(a.datecreated)) sorts the backups in descending order of creation date, meaning the newest backups are at the beginning of the array.

Slicing:
slice(5) removes the first 5 elements of the sorted array and returns the rest.  Since the newest backups are at the beginning, this line is effectively keeping backups older than the 5th newest and deleting the 5 most recent ones, including the one just created!

Fix:
slice(0, fileData.length - 5): This slices the array from the beginning (index 0) up to, but not including, the calculated index.